refactor(front): move isProviderWhitelisted to lib/api/assistant/models
