+++ front/lib/api/assistant/global_agents/configurations/dust/sidekick.ts
@@ -6,10 +6,10 @@ import type {
   PrefetchedDataSourcesType,
 } from "@app/lib/api/assistant/global_agents/tools";
 import { dummyModelConfiguration } from "@app/lib/api/assistant/global_agents/utils";
+import { isProviderWhitelisted } from "@app/lib/api/assistant/models";
 import {
   getLargeWhitelistedModel,
   getSmallWhitelistedModel,
-  isProviderWhitelisted,
 } from "@app/lib/assistant";
 import type { Authenticator } from "@app/lib/auth";
 import type {
