+++ front/lib/api/llm/getImageGenerationLLM.test.ts
@@ -14,7 +14,7 @@ vi.mock("@app/lib/api/provider_credentials", () => ({
   getLlmCredentials: mockGetLlmCredentials,
 }));
 
-vi.mock("@app/lib/assistant", () => ({
+vi.mock("@app/lib/api/assistant/models", () => ({
   isProviderWhitelisted: mockIsProviderWhitelisted,
 }));
